refactor: move pallet_revive deps into ink_sandbox
